keep renaming

+++ Unificado/utils.hs
@@ -19,17 +19,18 @@ getR (Right x) = x
 ordensindups :: (Ord a) => [a] -> [a]
 ordensindups = map head . group . sort
 
-maximo :: [(b,Int)] -> (b,Int) -> b
+maximo :: [(b,Double)] -> (b,Double) -> b
 maximo [] y                   = fst y
 maximo (x:xs) y 
            | (snd x) > (snd y) = maximo xs x
            | otherwise         = maximo xs y
 
-ocurrencia :: (Eq a) => [a] -> a -> Int
+ocurrencia :: (Eq a) => [a] -> a -> Double
 ocurrencia [] a      = 0
 ocurrencia (x:xs) a
            | a == x    = 1 + ocurrencia xs a
            | otherwise = ocurrencia xs a
+
 
 elimina _ []                 = []
 elimina x (y:ys) | x == y    = elimina x ys
@@ -50,31 +51,30 @@ valorObjetivo = snd.snd
 valores :: Atributo -> [Ejemplo] -> [ValorAtrib]
 valores _ [] = []
 valores atributo (e:ejemplos) =
-        [ snd x | x <- fst e, fst x == atributo ]
-        ++
-        valores atributo ejemplos
+    [ snd x | x <- fst e, fst x == atributo ]
+    ++
+    valores atributo ejemplos
 
 valorAtributo :: Ejemplo -> Atributo -> ValorAtrib
 valorAtributo ejemplo atributo =
-       snd $ head (filter (\x -> fst x == atributo) (fst ejemplo))
+    snd $ head (filter (\x -> fst x == atributo) (fst ejemplo))
        
 evaluarD :: [Ejemplo] -> String -> Discreto -> [Ejemplo]
 evaluarD ejemplos valor atributo =
-           let atrib = Left atributo 
-           in [x | x <- ejemplos, (getL $ valorAtributo x atrib) == valor ]
+    let atrib = Left atributo 
+    in [x | x <- ejemplos, (getL $ valorAtributo x atrib) == valor ]
 
 evaluarC :: [Ejemplo] -> String -> Continuo -> [Ejemplo]
 evaluarC ejemplos valor atributo =
-           let u = fromJust $ umbral atributo
-               atrib = (Right atributo)
-           in
-           if valor == "<="
-           then filter (\x -> (getR$valorAtributo x atrib) <= u) ejemplos
-           else if valor == ">"
-           then filter (\x -> (getR$valorAtributo x atrib) > u) ejemplos
-           else []
+    let u = fromJust $ umbral atributo
+        atrib = (Right atributo)
+    in if valor == "<="
+    then filter (\x -> (getR$valorAtributo x atrib) <= u) ejemplos
+    else if valor == ">"
+    then filter (\x -> (getR$valorAtributo x atrib) > u) ejemplos
+    else []
 
 evaluar :: [Ejemplo]  -> Atributo -> String -> [Ejemplo]
 evaluar ejemplos atributo valor =
-         either (evaluarD ejemplos valor) (evaluarC ejemplos valor) atributo
+    either (evaluarD ejemplos valor) (evaluarC ejemplos valor) atributo
